+++ Source/Foundation/bsfEngine/GUI/BsGUIHelper.h
@@ -86,7 +86,7 @@ namespace bs
 		 * @param	fontSize		Size of individual characters in the font, in points.
 		 * @return					Width/height required to display the text, in pixels.
 		 */
-		static Vector2I CalculateTextBounds(const String& text, const HFont& font, u32 fontSize);
+		static Vector2I CalculateTextBounds(const String& text, const HFont& font, float fontSize);
 	};
 
 	/** @} */
